+++ src/test/java/com/example/demo/builder/BuilderSupporter.java
@@ -1,14 +1,13 @@
 package com.example.demo.builder;
 
-import com.example.demo.infra.board.Repository.BoardJpaRepository;
-import com.example.demo.infra.board.Repository.LikeRepository;
-import com.example.demo.domain.comment.repository.BoardCommentRepository;
-import com.example.demo.domain.comment.repository.RecruitmentBoardCommentRepository;
 import com.example.demo.domain.notification.repository.NotificationRepository;
 import com.example.demo.domain.recruitment_board.repository.RecruitmentBoardRepository;
 import com.example.demo.domain.user.repository.UserJpaRepository;
+import com.example.demo.infra.board.Repository.BoardJpaRepository;
+import com.example.demo.infra.board.Repository.LikeRepository;
 import com.example.demo.infra.board.category.repository.CategoryJpaRepository;
-
+import com.example.demo.infra.comment.repository.jpa.BoardCommentJpaRepository;
+import com.example.demo.infra.comment.repository.jpa.RecruitmentBoardCommentJpaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -28,10 +27,10 @@ public class BuilderSupporter {
     private RecruitmentBoardRepository recruitmentBoardRepository;
 
     @Autowired
-    private BoardCommentRepository boardCommentRepository;
+    private BoardCommentJpaRepository boardCommentJpaRepository;
 
     @Autowired
-    private RecruitmentBoardCommentRepository recruitmentBoardCommentRepository;
+    private RecruitmentBoardCommentJpaRepository recruitmentBoardCommentJpaRepository;
 
     @Autowired
     private NotificationRepository notificationRepository;
@@ -56,18 +55,17 @@ public class BuilderSupporter {
         return recruitmentBoardRepository;
     }
 
-    public BoardCommentRepository boardCommentRepository() {
-        return boardCommentRepository;
+    public BoardCommentJpaRepository boardCommentRepository() {
+        return boardCommentJpaRepository;
     }
 
-    public RecruitmentBoardCommentRepository recruitmentBoardCommentRepository() {
-        return recruitmentBoardCommentRepository;
+    public RecruitmentBoardCommentJpaRepository recruitmentBoardCommentRepository() {
+        return recruitmentBoardCommentJpaRepository;
     }
 
     public NotificationRepository notificationRepository() {
         return notificationRepository;
     }
-
 
 
 }
