+++ src/test/java/com/example/demo/builder/TestFixtureBuilder.java
@@ -1,15 +1,12 @@
 package com.example.demo.builder;
 
-import java.util.List;
-
-import com.example.demo.infra.board.category.entity.Category;
-import com.example.demo.infra.board.entity.Board;
-import com.example.demo.infra.board.entity.Like;
-import com.example.demo.domain.comment.domain.entity.BoardComment;
-import com.example.demo.domain.comment.domain.entity.RecruitmentBoardComment;
 import com.example.demo.domain.notification.domain.entity.Notification;
 import com.example.demo.domain.recruitment_board.domain.entity.RecruitmentBoard;
 import com.example.demo.domain.user.domain.User;
+import com.example.demo.infra.board.entity.Board;
+import com.example.demo.infra.board.entity.Like;
+import com.example.demo.infra.comment.entity.BoardComment;
+import com.example.demo.infra.comment.entity.RecruitmentBoardComment;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
