+++ src/test/java/com/example/demo/fixture/comment/CommentFixtures.java
@@ -1,10 +1,10 @@
 package com.example.demo.fixture.comment;
 
-import com.example.demo.infra.board.entity.Board;
-import com.example.demo.domain.comment.domain.entity.BoardComment;
-import com.example.demo.domain.comment.domain.entity.RecruitmentBoardComment;
 import com.example.demo.domain.recruitment_board.domain.entity.RecruitmentBoard;
 import com.example.demo.domain.user.domain.User;
+import com.example.demo.infra.board.entity.Board;
+import com.example.demo.infra.comment.entity.BoardComment;
+import com.example.demo.infra.comment.entity.RecruitmentBoardComment;
 
 public class CommentFixtures {
     private static final String GENERAL_CONTENT = "content";
